requirements.txt
google-api-python-client~=2.112.0
google-auth-httplib2~=0.2.0
google-auth-oauthlib~=1.2.0
